cambios en los index

+++ ProyectoU3_Horarios/Codigo/php/usuarios-index.php
@@ -24,7 +24,7 @@
     <div class="container-fluid">
         <h1>Detalles de usuarios</h1>
         <div class="d-flex justify-content-end align-items-center mb-5">
-            <a href="usuarios-create.php" class="btn btn-success mr-3"><i class='bx bx-sm bx-plus'></i> Nuevo registro</a>
+            <a <?php if (!in_array("crear", $privilegios)){echo 'style="display: none;"';}?> href="usuarios-create.php" class="btn btn-success mr-3"><i class='bx bx-sm bx-plus'></i> Nuevo registro</a>
             <a href="usuarios-index.php" class="btn btn-info mr-3">Actualizar</a>
             <a href="index.php" class="btn btn-secondary"><i class='bx bx-sm bx-arrow-back'></i> Atrás</a>
         </div>
@@ -157,9 +157,24 @@
                     echo "<td>********</td>";
                     echo "<td>" . htmlspecialchars($row['id_perfil']) . "</td>";
                     echo "<td>";
-                    echo "<a href='usuarios-read.php?id_usuario=" . $row['id_usuario'] . "' title='Ver Registro' data-toggle='tooltip'><i class='far fa-eye'></i></a>";
-                    echo "<a href='usuarios-update.php?id_usuario=" . $row['id_usuario'] . "' title='Actualizar Registro' data-toggle='tooltip'><i class='far fa-edit'></i></a>";
-                    echo "<a href='usuarios-delete.php?id_usuario=" . $row['id_usuario'] . "' title='Eliminar Registro' data-toggle='tooltip'><i class='far fa-trash-alt'></i></a>";
+                    if (!in_array("seleccionar", $privilegios)){
+                        echo 'style="display: none;"';
+                    } else {
+                        echo "<a href='usuarios-read.php?id_usuario=" . $row['id_usuario'] . "' title='Ver Registro' data-toggle='tooltip'><i class='far fa-eye'></i></a>";
+                    }
+
+                    if (!in_array("editar", $privilegios)){
+                        echo 'style="display: none;"';
+                    } else {
+                        echo "<a href='usuarios-update.php?id_usuario=" . $row['id_usuario'] . "' title='Actualizar Registro' data-toggle='tooltip'><i class='far fa-edit'></i></a>";
+                    }
+
+                    if (!in_array("eliminar", $privilegios)){
+                        echo 'style="display: none;"';
+                    } else {
+                        echo "<a href='usuarios-delete.php?id_usuario=" . $row['id_usuario'] . "' title='Eliminar Registro' data-toggle='tooltip'><i class='far fa-trash-alt'></i></a>";
+                    }
+
                     echo "</td>";
                     echo "</tr>";
                 }
